Rename the folder and initialize the Contao framework in the run() method
